inbound-calling/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
livekit
asyncpg
